% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassGLM.R
\name{mfx.misclassGLM}
\alias{mfx.misclassGLM}
\title{Compute Marginal Effects for \code{misclassGLM} Fits}
\usage{
mfx.misclassGLM(w, x.mean = TRUE, rev.dum = TRUE, digits = 3, ...)
}
\arguments{
\item{w}{a fitted object of class inheriting from 'misclassGLM'.}

\item{x.mean}{logical, if true computes marginal effects at mean, otherwise average marginal
effects.}

\item{rev.dum}{logical, if true, computes differential effects for switch from 0 to 1.}

\item{digits}{number of digits to be presented in output.}

\item{\dots}{further arguments passed to or from other functions.}
}
\description{
Obtain marginal Effects.
}
\seealso{
\code{\link{misclassGLM}}
}
