% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassMLogit.R
\name{mfx.misclassMlogit}
\alias{mfx.misclassMlogit}
\title{Compute Marginal Effects for 'misclassMlogit' Fits}
\usage{
mfx.misclassMlogit(w, x.mean = TRUE, rev.dum = TRUE, outcome = 2,
  baseoutcome = 1, digits = 3, ...)
}
\arguments{
\item{w}{a fitted object of class inheriting from 'misclassMlogit'.}

\item{x.mean}{logical, if true computes marginal effects at mean, otherwise average marginal effects.}

\item{rev.dum}{logical, if true, computes differential effects for switch from 0 to 1.}

\item{outcome}{for which the ME should be computed.}

\item{baseoutcome}{base outcome, e.g. reference class of the model.}

\item{digits}{number of digits to be presented in output.}

\item{\dots}{further arguments passed to or from other functions.}
}
\description{
Obtain marginal effects.
}
\seealso{
\code{\link{misclassMlogit}}
}

