% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassGLM-package.R, R/misclassGLM.R
\docType{package}
\name{misclassGLM}
\alias{misclassGLM}
\alias{misclassGLM-package}
\title{misclassGLM}
\usage{
misclassGLM(Y, X, setM, P, na.action = na.omit, family = gaussian(link =
  "identity"), control = list(), par = NULL, x = FALSE, robust = FALSE)
}
\arguments{
\item{Y}{a vector of integers or numerics. This is the dependent variable.}

\item{X}{a matrix containing the independent variables}

\item{setM}{matrix, rows containing potential patterns for a misclassified (latent) covariate M in any coding for a
categorical independent variable, e.g. dummy coding.}

\item{P}{probabilities corresponding to each of the potential pattern conditional on the other covariates denoted in x.}

\item{na.action}{how to treat NAs}

\item{family}{a description of the error distribution and link function to be used in the model.
This can be a character string naming a family function, a family function or the result of a call to a
family function. (See \code{\link{family}} for details of family functions.)}

\item{control}{options for the optimization procedure (see \code{\link{optim}}, \code{\link{ucminf}} for options and details).}

\item{par}{(optional) starting parameter vector}

\item{x}{logical, add covariates matrix to result?}

\item{robust}{logical, if true the computed asymptotic standard errors are replaced by their robust counterparts}
}
\description{
Estimates models that extend the standard GLM to take
misclassification into account. The models require side information from a secondary data set
on the misclassification process, i.e. some sort of misclassification
probabilities conditional on some common covariates.
A detailed description of the algorithm can be found in
Dlugosz, Mammen and Wilke (2015) \url{http://www.zew.de/PU70410}.

\code{misclassGLM} computes estimator for a GLM with a misclassified covariate
using additional side information on the misclassification process
}
\details{
The two main functions are \code{\link{misclassGLM}} and \code{\link{misclassMlogit}}.
}
\examples{
## simulate data
\donttest{
data <- simulate_GLM_dataset()
}

## estimate model without misclassification error
\donttest{
summary(lm(Y ~ X + M2, data))
}

## estimate model with misclassification error
\donttest{
summary(lm(Y ~ X + M, data))
}

## estimate misclassification probabilities
\donttest{
Pmodel <- glm(M2 ~ M + X, data = data, family = binomial("logit"))
summary(Pmodel)
}

## construct a-posteriori probabilities from Pmodel
\donttest{
P <- predict(Pmodel, newdata = data, type = "response")
P <- cbind(1 - P, P)
dimnames(P)[[2]] <- c("M0", "M1") ## speaking names
}

## estimate misclassGLM
\donttest{
est <- misclassGLM(Y = data$Y,
                   X = as.matrix(data[, 2, drop = FALSE]),
                   setM = matrix(c(0, 1), nrow = 2),
                   P = P)
summary(est)
}

## and bootstrapping the results from dataset
\dontrun{
  summary(boot.misclassGLM(est,
                           Y = data$Y,
                           X = data.matrix(data[, 2, drop = FALSE]),
                           Pmodel = Pmodel,
                           PX = data,
                           repetitions = 100))
}

}

